Bug 1304876: Fix 64-bit build failure in IPC TestDataStructures test; r=billm
MozReview-Commit-ID: G0b0mvV8gus

+++ ipc/ipdl/test/cxx/TestDataStructures.cpp
@@ -166,7 +166,7 @@ TestDataStructuresParent::RecvTest7_0(const ActorWrapper& i1,
     o1->actorParent() = mKids[0];
     // malicious behavior
     o1->actorChild() =
-        reinterpret_cast<PTestDataStructuresSubChild*>(0xdeadbeef);
+        reinterpret_cast<PTestDataStructuresSubChild*>(uintptr_t(0xdeadbeef));
     return true;
 }
